refactor: layoutディレクトリをLayoutディレクトリにして関係ないコンポーネントはcommonに移動した
