+++ src/resources/ts/ChatBookmarks/ChatBookmarks.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { List } from '@mui/material';
-import { main } from '@/ts/layout/color';
+import { main } from '@/ts/common/color';
 import { ChatBookmarkListHeader } from '@/ts/ChatBookmarks/ChatBookmarkListHeader';
 import { ChatBookmarkList } from '@/ts/ChatBookmarks/ChatBookmarkList';
