+++ src/resources/ts/Chats/ChatCreateButton.tsx
@@ -1,7 +1,7 @@
 import { IconButton } from '@mui/material';
 import React from 'react';
 import AddCircleIcon from '@mui/icons-material/AddCircle';
-import { sub } from '@/ts/layout/color';
+import { sub } from '@/ts/common/color';
 
 type ChatCreateButtonProps = {
     openDialog: () => void;
