+++ src/resources/ts/common/ChatBookmarkRegisterButton.tsx
@@ -3,7 +3,7 @@ import { registerChatBookmark } from '@/ts/http/chatBookmarks';
 import { IconButton } from '@mui/material';
 import BookmarkBorderIcon from '@mui/icons-material/BookmarkBorder';
 import { ChatBookmarkContext } from '@/ts/ChatBookmarks/ChatBookmarkProvider';
-import { sub } from '@/ts/layout/color';
+import { sub } from '@/ts/common/color';
 import BookmarkIcon from '@mui/icons-material/Bookmark';
 
 type ChatBookmarkButtonProps = {
