+++ src/resources/ts/common/ErrorAlert.tsx
