+++ src/resources/ts/common/color.ts
